+++ .idea/gradle.xml
@@ -13,6 +13,7 @@
             <option value="$PROJECT_DIR$/app" />
             <option value="$PROJECT_DIR$/challenge_home" />
             <option value="$PROJECT_DIR$/challenge_model" />
+            <option value="$PROJECT_DIR$/challenge_produto" />
             <option value="$PROJECT_DIR$/coreui" />
             <option value="$PROJECT_DIR$/mylibrary" />
           </set>
