+++ challenge_produto/build.gradle
@@ -56,7 +56,6 @@ dependencies {
     implementation("io.coil-kt:coil-compose:2.0.0-rc03")
     implementation project(path: ':coreui')
     implementation project(path: ':challenge_model')
-    implementation project(path: ':challenge_produto')
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
