+++ challenge_produto/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.amedigital.challenge_home">
+    package="com.amedigital.challenge_produto">
 </manifest>
