+++ challenge_produto/src/main/java/com/amedigital/challenge_produto/ListaProdutos.kt
@@ -1,19 +1,15 @@
-package com.amedigital.challenge_home
+package com.amedigital.challenge_produto
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.horizontalScroll
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.rememberScrollState
 import androidx.compose.material.Divider
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
@@ -21,13 +17,11 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
-import com.amedigital.challenge_model.Categoria
 import com.amedigital.challenge_model.Produto
 import com.amedigital.coreui.extensions.toMoneyBR
 import com.amedigital.coreui.theme.ChallengeColors
@@ -63,7 +57,9 @@ fun ListaProdutos(
                     Spacer(modifier = Modifier.width(8.dp))
                     Column() {
                         Text(
-                            modifier = Modifier.width(constraints.maxWidth - 100.dp),
+                            modifier = Modifier
+                                .width(constraints.maxWidth - 100.dp)
+                                .padding(end = 8.dp),
                             text = produto.descricao, maxLines = 3,
                             overflow = TextOverflow.Ellipsis
                         )
@@ -85,7 +81,6 @@ fun ListaProdutos(
                         }
                     }
                 }
-
                 Image(
                     modifier = Modifier.size(16.dp),
                     painter = painterResource(R.drawable.disclosure_indicator),
