+++ settings.gradle
@@ -18,3 +18,4 @@ include ':challenge_home'
 include ':mylibrary'
 include ':coreui'
 include ':challenge_model'
+include ':challenge_produto'
